# workers/transcriptor-legal/openapi.yaml
openapi: 3.1.0
info:
  title: Transcriptor Legal
  version: 2.0.0
  description: >-
    Recibe un archivo del usuario y recupera segmentos literales en español desde
    Cloudflare Stream. Conserva job_id y job_token; utiliza este token en todas las
    consultas posteriores como parámetro query. completed indica que captions está
    listo: recupera páginas de hasta 90.000 bytes y caracteres hasta next_offset null.
    No conviertas mensajes de diagnóstico en
    transcripciones. El borrado necesita autorización expresa y no verifica Word.
servers:
  - url: https://transcriptor-legal.yoshiman1989.workers.dev
security:
  - ActionKey: []
paths:
  /jobs:
    post:
      operationId: crearTranscripcion
      summary: Recibir un archivo de audio o video
      x-openai-isConsequential: true
      description: >-
        Espera exactamente un archivo aportado por el usuario. Mantén el mismo
        request_id al recuperar una respuesta perdida; no vuelvas a subir el archivo
        si la API devuelve recovery_required. No anuncia Whisper: usa captions de Stream.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [openaiFileIdRefs]
              properties:
                openaiFileIdRefs:
                  type: array
                  minItems: 1
                  maxItems: 1
                  description: Un archivo de audio o video cargado por el usuario para transcripción literal.
                  items:
                    type: string
                request_id:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]{1,128}$'
                  description: Identificador único que debe mantenerse al repetir la misma solicitud.
      responses:
        '200':
          description: Identificador confirmado; conservar el token sin divulgarlo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedJob'
        '409':
          $ref: '#/components/responses/ControlledError'
        '502':
          $ref: '#/components/responses/ControlledError'
  /jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobId'
      - $ref: '#/components/parameters/JobToken'
    get:
      operationId: consultarEstadoTranscripcion
      summary: Consultar procesamiento del video y subtítulos
      x-openai-isConsequential: true
      description: >-
        Esta consulta puede iniciar captions cuando el video esté listo. completed
        no incluye texto completo; recupera las páginas de transcript. No interpretes
        video_status.pctComplete como progreso de captions.
      responses:
        '200':
          description: Estado real de video/captions o error terminal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '400':
          $ref: '#/components/responses/ControlledError'
        '404':
          $ref: '#/components/responses/ControlledError'
        '502':
          $ref: '#/components/responses/ControlledError'
    delete:
      operationId: eliminarVideo
      summary: Eliminar el video remoto autorizado
      x-openai-isConsequential: true
      description: >-
        Ejecuta solamente si el usuario autoriza la eliminación. Este endpoint no
        acredita generación o revisión de Word; Lexiapp añade ese control en la fase 4.
      responses:
        '200':
          description: Eliminación confirmada
          content:
            application/json:
              schema:
                type: object
                required: [job_id, deleted]
                properties:
                  job_id: { type: string }
                  deleted: { type: boolean, const: true }
        '400':
          $ref: '#/components/responses/ControlledError'
        '404':
          $ref: '#/components/responses/ControlledError'
  /jobs/{job_id}/transcript:
    get:
      operationId: recuperarSegmentosTranscripcion
      summary: Recuperar una página literal con tiempos y hablantes identificados
      description: >-
        Recupera cues completos hasta 90.000 bytes y caracteres de JSON. Continúa
        con next_offset hasta null; si un cue no cabe, devuelve error sin recortarlo.
      x-openai-isConsequential: false
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/JobToken'
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        '200':
          description: Página verificada; continuar desde next_offset y conservar transcript_hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptPage'
        '400':
          $ref: '#/components/responses/ControlledError'
        '404':
          $ref: '#/components/responses/ControlledError'
        '409':
          $ref: '#/components/responses/ControlledError'
        '422':
          $ref: '#/components/responses/ControlledError'
        '502':
          $ref: '#/components/responses/ControlledError'
components:
  securitySchemes:
    ActionKey:
      type: http
      scheme: bearer
  parameters:
    JobId:
      in: path
      name: job_id
      required: true
      schema: { type: string, pattern: '^[a-zA-Z0-9_-]{1,128}$' }
    JobToken:
      in: query
      name: job_token
      required: true
      description: Token devuelto al crear este trabajo; conservarlo sin publicarlo.
      schema: { type: string }
  responses:
    ControlledError:
      description: Error controlado sin diagnósticos internos del proveedor
      content:
        application/json:
          schema:
            type: object
            required: [error, code]
            properties:
              error: { type: string }
              code: { type: string }
  schemas:
    VideoStatus:
      type: object
      properties:
        state: { type: string }
        pctComplete: { type: number, minimum: 0, maximum: 100 }
    CreatedJob:
      type: object
      required: [job_id, job_token, request_id, status]
      properties:
        job_id: { type: string }
        job_token: { type: string }
        request_id: { type: string }
        filename: { type: string }
        status:
          type: string
          enum: [processing_video, video_ready, error]
        ready_to_stream: { type: boolean }
        video_status:
          $ref: '#/components/schemas/VideoStatus'
        scheduled_deletion: { type: [string, 'null'], format: date-time }
        scheduled_deletion_requested: { type: string, format: date-time }
    JobStatus:
      type: object
      required: [job_id, language, status]
      properties:
        job_id: { type: string }
        language: { type: string, const: es }
        status:
          type: string
          enum: [uploading, processing_video, generating_transcript, completed, error]
        video_status:
          $ref: '#/components/schemas/VideoStatus'
        ready_to_stream: { type: boolean }
        duration_seconds: { type: [number, 'null'] }
        total_segments: { type: integer, minimum: 1 }
        transcript_hash: { type: string, pattern: '^[a-f0-9]{64}$' }
        scheduled_deletion: { type: [string, 'null'], format: date-time }
        error: { type: string }
        code: { type: string }
    Segment:
      type: object
      required: [index, start, end, text]
      properties:
        index: { type: integer, minimum: 0 }
        start: { type: number, minimum: 0 }
        end: { type: number, minimum: 0 }
        text: { type: string }
        speaker:
          type: string
          description: Solo se incluye si WebVTT lo identifica explícitamente.
    TranscriptPage:
      type: object
      required: [job_id, status, language, total_segments, offset, limit, next_offset, segments, transcript_hash]
      properties:
        job_id: { type: string }
        status: { type: string, const: completed }
        language: { type: string, const: es }
        total_segments: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 200 }
        next_offset: { type: [integer, 'null'], minimum: 0 }
        transcript_hash: { type: string, pattern: '^[a-f0-9]{64}$' }
        segments:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/Segment'
        text: { type: string, description: Texto de esta página solamente; no representa el documento completo. }